requirements.txt
fastapi
uvicorn
playwright
pydantic
python-socketio[asyncio]
aiohttp
